Create README.md
The primary objective of this analysis in the initial part (task-1) of the project was to design a communication link for AM audio transmission. Firstly, we will trigger a message signal , then modulate it with a carrier, and the add noise to the modulated signal using an AWGN (Additive White Gaussian Noise) channel. Then we will need to demodulate this modulated signal in order to recover our original message signal.